Added files via upload
This PowerShell script enumerates domain joined servers in Active Directory and accesses personal certificate stores to generate a report for expiring certificates. This does not require PSRemoting to be enabled as it uses .NET. 
This script must be ran by an account with privileges necessary to connect the local certificate store on the servers.